+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
  
 ## [Unreleased]
+### Added
+- Agent mode that can be used to indicate an app is running in the background
 ### Changed
 - **BREAKING**: Renamed property `debugMode` to `debug` in `Agent` constructor. \
 `debug` will also accept a callback function that is called for all log messages.